requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0

# VLM Dependencies
transformers>=4.30.0
clip-openai>=1.0.1
timm>=0.9.0

# Data Processing
pandas>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Evaluation
scikit-learn>=1.3.0
scipy>=1.11.0

# Optional: GPU Acceleration
# onnxruntime-gpu>=1.15.0

# Development
jupyter>=1.0.0
ipython>=8.12.0
black>=23.3.0
flake8>=6.0.0
